feat(Onfido): AND-1231 Various bits of tidying up. Added GetStringFromResource.
