+++ buysell/src/test/java/piuk/blockchain/androidbuysell/services/CoinifyServiceTest.kt
@@ -2,7 +2,6 @@ package piuk.blockchain.androidbuysell.services
 
 import com.nhaarman.mockito_kotlin.whenever
 import com.squareup.moshi.Moshi
-import okhttp3.OkHttpClient
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import okhttp3.mockwebserver.RecordedRequest
@@ -13,6 +12,7 @@ import org.amshove.kluent.mock
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import piuk.blockchain.android.testutils.MockedRetrofitTest
 import piuk.blockchain.android.testutils.mockWebServerInit
 import piuk.blockchain.androidbuysell.api.PATH_COINFY_AUTH
 import piuk.blockchain.androidbuysell.api.PATH_COINFY_BANK_ACCOUNTS
@@ -52,9 +52,6 @@ import piuk.blockchain.androidbuysell.models.coinify.exceptions.CoinifyApiExcept
 import piuk.blockchain.androidbuysell.models.coinify.exceptions.CoinifyErrorCodes
 import piuk.blockchain.androidcore.data.api.EnvironmentConfig
 import piuk.blockchain.androidcore.data.rxjava.RxBus
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import retrofit2.converter.moshi.MoshiConverterFactory
 
 class CoinifyServiceTest {
 
@@ -70,8 +67,6 @@ class CoinifyServiceTest {
         .add(GrantTypeAdapter())
         .add(BuyFrequencyAdapter())
         .build()
-    private val moshiConverterFactory = MoshiConverterFactory.create(moshi)
-    private val rxJava2CallAdapterFactory = RxJava2CallAdapterFactory.create()
     private val environmentConfig: EnvironmentConfig = mock()
 
     private val server = MockWebServer()
@@ -81,18 +76,9 @@ class CoinifyServiceTest {
 
     @Before
     fun setUp() {
-        val okHttpClient = OkHttpClient.Builder()
-            .build()
-        val retrofit = Retrofit.Builder()
-            .client(okHttpClient)
-            .baseUrl(server.url("/").toString())
-            .addConverterFactory(moshiConverterFactory)
-            .addCallAdapterFactory(rxJava2CallAdapterFactory)
-            .build()
-
         whenever(environmentConfig.coinifyUrl).thenReturn("")
 
-        subject = CoinifyService(environmentConfig, retrofit, rxBus)
+        subject = CoinifyService(environmentConfig, MockedRetrofitTest(moshi, server).retrofit, rxBus)
     }
 
     @Test
