--- kyc/src/test/java/piuk.blockchain.kyc/ExampleTest.kt
@@ -1,12 +1,0 @@
-package piuk.blockchain.kyc
-
-import org.amshove.kluent.`should be`
-import org.junit.Test
-
-class ExampleTest {
-
-    @Test
-    fun `I pass`() {
-        1 `should be` 1
-    }
-}
