+++ kyc/src/test/resources/com/blockchain/kyc/services/onfido/CheckResponse.json
@@ -1,0 +1,36 @@
+{
+  "id": "8546921-123123-123123",
+  "created_at": "2014-05-23T13:50:33Z",
+  "href": "/v2/applicants/1030303-123123-123123/checks/8546921-123123-123123",
+  "type": "standard",
+  "status": "awaiting_applicant",
+  "result": null,
+  "results_uri": "https://onfido.com/dashboard/information_requests/1234",
+  "redirect_uri": null,
+  "reports": [
+    {
+      "id": "6951786-123123-422221",
+      "name": "identity",
+      "created_at": "2014-05-23T13:50:33Z",
+      "status": "awaiting_data",
+      "result": null,
+      "href": "/v2/checks/8546921-123123-123123/reports/6951786-123123-422221",
+      "breakdown": {},
+      "properties": {}
+    },
+    {
+      "id": "6951786-123123-316712",
+      "name": "document",
+      "created_at": "2014-05-23T13:50:33Z",
+      "status": "awaiting_data",
+      "result": null,
+      "href": "/v2/checks/8546921-123123-123123/reports/6951786-123123-316712",
+      "breakdown": {},
+      "properties": {}
+    }
+  ],
+  "tags": [
+    "My tag",
+    "Another tag"
+  ]
+}
