+++ kyc/src/test/resources/com/blockchain/kyc/services/onfido/CreateApplicantResponse.json
@@ -1,0 +1,23 @@
+{
+  "id": "6a29732d-4561-4760-a2e3-a244ad324ba2",
+  "created_at": "2018-07-25T13:30:55Z",
+  "sandbox": true,
+  "title": null,
+  "first_name": "Theresa",
+  "middle_name": null,
+  "last_name": "May",
+  "email": null,
+  "gender": null,
+  "dob": null,
+  "telephone": null,
+  "mobile": null,
+  "nationality": null,
+  "country": "gbr",
+  "href": "/v2/applicants/6a29732d-4561-4760-a2e3-a244ad324ba2",
+  "mothers_maiden_name": null,
+  "country_of_birth": null,
+  "town_of_birth": null,
+  "previous_last_name": null,
+  "id_numbers": [],
+  "addresses": []
+}
