+++ testutils/build.gradle
@@ -23,6 +23,10 @@ dependencies {
     api Libraries.rxAndroid
     // Networking
     api Libraries.okHttpMock
+    api Libraries.retrofit
+    api Libraries.moshi
+    api Libraries.retrofitRxMoshiConverter
+    api Libraries.retrofitRxJavaAdapter
 
     // Unit Test dependencies
     api Libraries.junit
