+++ testutils/src/main/java/piuk/blockchain/android/testutils/GetStringFromResource.kt
@@ -1,0 +1,8 @@
+package piuk.blockchain.android.testutils
+
+fun Any.getStringFromResource(filePath: String): String =
+    this::class.java.getResource("/$filePath").readText()
+
+fun String.asResource(work: (String) -> Unit) {
+    work(getStringFromResource(this))
+}
