+++ testutils/src/test/resources/GetStringFromResourceTestFile.txt
@@ -1,0 +1,1 @@
+Test string
